server/back/requirements.txt
fastapi
uvicorn[standard]
passlib
PyJWT
datetime
motor
bcrypt
python-multipart